+++ src/Writer/Stream.php
@@ -79,6 +79,10 @@ class Stream extends AbstractWriter
             $this->stream = $streamOrUrl;
         } else {
             ErrorHandler::start();
+            if (isset($filePermissions) && !file_exists($streamOrUrl) && is_writable(dirname($streamOrUrl))) {
+                touch($streamOrUrl);
+                chmod($streamOrUrl, $filePermissions);
+            }
             $this->stream = fopen($streamOrUrl, $mode, false);
             $error = ErrorHandler::stop();
             if (!$this->stream) {
@@ -87,13 +91,6 @@ class Stream extends AbstractWriter
                     $streamOrUrl,
                     $mode
                 ), 0, $error);
-            }
-            if (null !== $filePermissions && !chmod($streamOrUrl, $filePermissions)) {
-                throw new Exception\RuntimeException(sprintf(
-                    'Could not set the mode "%o" for the log file "%s"',
-                    $filePermissions,
-                    $streamOrUrl
-                ));
             }
         }